+++ runtime/src/app/app.ts
@@ -1,15 +1,20 @@
 import { writable } from 'svelte/store';
 import App from '@sapper/internal/App.svelte';
-import { root_comp, ErrorComponent, ignore, components, routes } from '@sapper/internal/manifest-client';
+import { Query } from '@sapper/internal/shared';
 import {
+	DOMComponentLoader,
+	DOMComponentModule,
+	ErrorComponent,
+	ignore,
+	components,
+	root_comp,
+	routes
+} from '@sapper/internal/manifest-client';
+import {
+	HydratedTarget,
 	Target,
 	ScrollPosition,
-	Component,
 	Redirect,
-	ComponentLoader,
-	ComponentConstructor,
-	Route,
-	Query,
 	Page
 } from './types';
 import goto from './goto';
@@ -20,7 +25,7 @@ declare const __SAPPER__;
 export const initial_data = typeof __SAPPER__ !== 'undefined' && __SAPPER__;
 
 let ready = false;
-let root_component: Component;
+let root_component: InstanceType<typeof App>;
 let current_token: {};
 let root_preloaded: Promise<any>;
 let current_branch = [];
@@ -47,18 +52,22 @@ stores.session.subscribe(async value => {
 	const { redirect, props, branch } = await hydrate_target(target);
 	if (token !== current_token) return; // a secondary navigation happened while we were loading
 
-	await render(redirect, branch, props, target.page);
+	if (redirect) {
+		await goto(redirect.location, { replaceState: true });
+	} else {
+		await render(branch, props, target.page);
+	}
 });
 
 export let prefetching: {
 	href: string;
-	promise: Promise<{ redirect?: Redirect, data?: any }>;
+	promise: Promise<HydratedTarget>;
 } = null;
 export function set_prefetching(href, promise) {
 	prefetching = { href, promise };
 }
 
-export let target: Node;
+export let target: Element;
 export function set_target(element) {
 	target = element;
 }
@@ -86,7 +95,7 @@ export function extract_query(search: string) {
 	const query = Object.create(null);
 	if (search.length > 0) {
 		search.slice(1).split('&').forEach(searchParam => {
-			let [, key, value = ''] = /([^=]*)(?:=(.*))?/.exec(decodeURIComponent(searchParam.replace(/\+/g, ' ')));
+			const [, key, value = ''] = /([^=]*)(?:=(.*))?/.exec(decodeURIComponent(searchParam.replace(/\+/g, ' ')));
 			if (typeof query[key] === 'string') query[key] = [<string>query[key]];
 			if (typeof query[key] === 'object') {
 				(query[key] as string[]).push(value);
@@ -144,7 +153,7 @@ export function handle_error(url: URL) {
 	const { session, preloaded, status, error } = initial_data;
 
 	if (!root_preloaded) {
-		root_preloaded = preloaded && preloaded[0]
+		root_preloaded = preloaded && preloaded[0];
 	}
 
 	const props = {
@@ -163,9 +172,9 @@ export function handle_error(url: URL) {
 		},
 		segments: preloaded
 
-	}
+	};
 	const query = extract_query(search);
-	render(null, [], props, { host, path: pathname, query, params: {} });
+	render([], props, { host, path: pathname, query, params: {} });
 }
 
 export function scroll_state() {
@@ -200,11 +209,17 @@ export async function navigate(target: Target, id: number, noscroll?: boolean, h
 	prefetching = null;
 
 	const token = current_token = {};
-	const { redirect, props, branch } = await loaded;
+	const loaded_result = await loaded;
+	const { redirect } = loaded_result;
 	if (token !== current_token) return; // a secondary navigation happened while we were loading
 
-	await render(redirect, branch, props, target.page);
-	if (document.activeElement) document.activeElement.blur();
+	if (redirect) {
+		await goto(redirect.location, { replaceState: true });
+	} else {
+		const { props, branch } = loaded_result;
+		await render(branch, props, target.page);
+	}
+	if (document.activeElement && (document.activeElement instanceof HTMLElement)) document.activeElement.blur();
 
 	if (!noscroll) {
 		scroll_to_target(hash);
@@ -243,9 +258,7 @@ export function scroll_to_target(hash: string) {
 	if (scroll) scrollTo(scroll.x, scroll.y);
 }
 
-async function render(redirect: Redirect, branch: any[], props: any, page: Page) {
-	if (redirect) return goto(redirect.location, { replaceState: true });
-
+async function render(branch: any[], props: any, page: Page) {
 	stores.page.set(page);
 	stores.preloading.set(false);
 
@@ -292,11 +305,7 @@ function part_changed(i, segment, match, stringified_query) {
 	}
 }
 
-export async function hydrate_target(target: Target): Promise<{
-	redirect?: Redirect;
-	props?: any;
-	branch?: Array<{ Component: ComponentConstructor, preload: (page) => Promise<any>, segment: string }>;
-}> {
+export async function hydrate_target(target: Target): Promise<HydratedTarget> {
 	const { route, page } = target;
 	const segments = page.path.split('/').filter(Boolean);
 
@@ -306,7 +315,8 @@ export async function hydrate_target(target: Target): Promise<{
 
 	if (route === null && initial_data.hashbang) {
 		if (!page.path) {
-			return { redirect: { statusCode: 302, location: '#!/' } }
+			redirect = { statusCode: 302, location: '#!/' };
+			return { redirect, props, branch: [] };
 		}
 		props.error = { message: 'Page not found.' };
 		props.status = 404;
@@ -411,10 +421,7 @@ function load_css(chunk: string) {
 	});
 }
 
-export function load_component(component: ComponentLoader): Promise<{
-	default: ComponentConstructor,
-	preload?: (input: any) => any
-}> {
+export function load_component(component: DOMComponentLoader): Promise<DOMComponentModule> {
 	// TODO this is temporary — once placeholders are
 	// always rewritten, scratch the ternary
 	const promises: Array<Promise<any>> = (typeof component.css === 'string' ? [] : component.css.map(load_css));