+++ runtime/src/internal/App.svelte.d.ts
@@ -1,0 +1,8 @@
+import {
+  ComponentConstructor,
+  SSRComponent
+} from './shared';
+
+declare const App: ComponentConstructor & SSRComponent;
+
+export default App;