+++ runtime/src/server/middleware/index.ts
@@ -1,12 +1,11 @@
 import fs from 'fs';
 import path from 'path';
 import mime from 'mime/lite';
-import { build_dir, dev, manifest } from '@sapper/internal/manifest-server';
-import { Handler, Req, Res } from './types';
+import { Handler, Req, Res, build_dir, dev, manifest } from '@sapper/internal/manifest-server';
 import { get_server_route_handler } from './get_server_route_handler';
 import { get_page_handler } from './get_page_handler';
 
-type IgnoreValue = Array | RegExp | function | string;
+type IgnoreValue = IgnoreValue[] | RegExp | ((uri: string) => boolean) | string;
 
 export default function middleware(opts: {
 	session?: (req: Req, res: Res) => any,
@@ -109,7 +108,7 @@ export function serve({ prefix, pathname, cache_control }: {
 
 	const read = dev
 		? (file: string) => fs.readFileSync(path.join(build_dir, file))
-		: (file: string) => (cache.has(file) ? cache : cache.set(file, fs.readFileSync(path.join(build_dir, file)))).get(file)
+		: (file: string) => (cache.has(file) ? cache : cache.set(file, fs.readFileSync(path.join(build_dir, file)))).get(file);
 
 	return (req: Req, res: Res, next: () => void) => {
 		if (filter(req)) {
@@ -138,4 +137,4 @@ export function serve({ prefix, pathname, cache_control }: {
 	};
 }
 
-function noop(){}
+function noop() {}