+++ test/apps/errors/src/routes/preload-reject.svelte
@@ -1,6 +1,6 @@
 <script context="module">
 	export function preload() {
-		return Promise.reject(new Error('boom'))
+		return Promise.reject(new Error('boom'));
 	}
 </script>